refactor: move handleLegacySideEffects
